# Conflicts:
#	src/main/java/com/example/adressbook/AdressbookApplication.java
#	src/main/java/com/example/adressbook/controller/ContactController.java
#	src/main/java/com/example/adressbook/entity/Contact.java
#	src/main/java/com/example/adressbook/service/ContactService.java

+++ src/main/java/com/example/adressbook/repository/ContactRepository.java
@@ -1,7 +1,6 @@
 package com.example.adressbook.repository;
-
 import com.example.adressbook.entity.Contact;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ContactRepository extends MongoRepository<Contact,Integer> {
+public interface ContactRepository extends MongoRepository<Contact, Long> {
 }
